% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_determinant.R
\name{matrix_determinant}
\alias{matrix_determinant}
\title{Calculate the determinant of a square matrix.}
\usage{
matrix_determinant(mat)
}
\arguments{
\item{mat}{A square matrix.}
}
\value{
The determinant of the matrix.
}
\description{
This function takes a square matrix as input and computes its determinant.
}
\examples{
mat <- matrix(c(1, 2, 3, 4), nrow = 2)
matrix_determinant(mat)
}
