% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{matrix_eigen}
\alias{matrix_eigen}
\title{Calculate the eigenvalues and eigenvectors of a square matrix.}
\usage{
matrix_eigen(mat)
}
\arguments{
\item{mat}{A square matrix.}
}
\value{
A list with components 'values' (eigenvalues) and 'vectors' (eigenvectors).
}
\description{
This function takes a square matrix as input and computes its eigenvalues and eigenvectors.
}
\examples{
mat <- matrix(c(1, 2, 3, 4), nrow = 2)
matrix_eigen(mat)
}
