% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{matrix_frobenius_norm}
\alias{matrix_frobenius_norm}
\title{Calculate the Frobenius norm of a matrix.}
\usage{
matrix_frobenius_norm(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
The Frobenius norm of the matrix.
}
\description{
This function calculates the Frobenius norm of a matrix.
}
\examples{
mat <- matrix(c(1, 2, 3, 4), nrow = 2)
matrix_frobenius_norm(mat)
}
