% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R, R/matrix_inverse.R.R
\name{matrix_inverse}
\alias{matrix_inverse}
\title{Calculate the inverse of a square matrix.}
\usage{
matrix_inverse(mat)

matrix_inverse(mat)
}
\arguments{
\item{mat}{A square matrix.}
}
\value{
The inverse matrix if it exists; otherwise, an error is returned.

The inverse matrix if it exists; otherwise, an error is returned.
}
\description{
This function takes a square matrix as input and computes its inverse.

This function takes a square matrix as input and computes its inverse.
}
\examples{
mat <- matrix(c(1, 2, 3, 4), nrow = 2)
matrix_inverse(mat)
mat <- matrix(c(1, 2, 3, 4), nrow = 2)
matrix_inverse(mat)
}
