% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{matrix_multiply}
\alias{matrix_multiply}
\title{Multiply two matrices.}
\usage{
matrix_multiply(mat1, mat2)
}
\arguments{
\item{mat1}{First matrix.}

\item{mat2}{Second matrix.}
}
\value{
The product of the two matrices.
}
\description{
This function multiplies two matrices and returns the result.
}
\examples{
mat1 <- matrix(c(1, 2, 3, 4), nrow = 2)
mat2 <- matrix(c(2, 0, 1, 3), nrow = 2)
matrix_multiply(mat1, mat2)
}
