% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{matrix_power}
\alias{matrix_power}
\title{Calculate the matrix raised to a power.}
\usage{
matrix_power(mat, power)
}
\arguments{
\item{mat}{A square matrix.}

\item{power}{The power to which the matrix should be raised.}
}
\value{
The matrix raised to the power.
}
\description{
This function takes a square matrix and a power as input and computes the matrix raised to the power.
}
\examples{
mat <- matrix(c(1, 2, 3, 4), nrow = 2)
matrix_power(mat, 2)
}
