% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{matrix_rank}
\alias{matrix_rank}
\title{Calculate the rank of a matrix.}
\usage{
matrix_rank(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
The rank of the matrix.
}
\description{
This function calculates the rank of a matrix.
}
\examples{
mat <- matrix(c(1, 2, 3, 4), nrow = 2)
matrix_rank(mat)
}
