% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{matrix_spectral_norm}
\alias{matrix_spectral_norm}
\title{Calculate the spectral norm of a matrix.}
\usage{
matrix_spectral_norm(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
The spectral norm of the matrix.
}
\description{
This function calculates the spectral norm of a matrix.
}
\examples{
mat <- matrix(c(1, 2, 3, 4), nrow = 2)
matrix_spectral_norm(mat)
}
