% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{matrix_svd}
\alias{matrix_svd}
\title{Perform Singular Value Decomposition (SVD) on a matrix.}
\usage{
matrix_svd(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
A list with components 'u', 'd', and 'v' representing the SVD of the matrix.
}
\description{
This function takes a matrix as input and performs Singular Value Decomposition (SVD).
}
\examples{
mat <- matrix(c(1, 2, 3, 4), nrow = 2)
matrix_svd(mat)
}
