% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{matrix_transpose}
\alias{matrix_transpose}
\title{Calculate the transpose of a matrix.}
\usage{
matrix_transpose(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
The transpose of the input matrix.
}
\description{
This function calculates the transpose of a matrix.
}
\examples{
mat <- matrix(c(1, 2, 3, 4), nrow = 2)
matrix_transpose(mat)
}
